Hmmm... getchar() and getch() and getc() don't seem to grab the character without getting a carriage return or "\n" but getche from ncurses (or is it from curses) grabs a worn out buffer and throws a segmentation fault because it doesn't seem to clear the buffer and as of yet I haven't figured out how to clear the buffer satisfactorily either.
I'd like to interactively be reading input from the user and be able to do things
in between their keypresses, but for now I think I'm just going to write functions
that call getline and refine them later when I can.  Would like to get to parsing
rooms and then writing them to file so the game can be built.  Changing things
to be more aesthetic should be a later feature built when time can be spent
on those individual issues